requirements.txt
python-telegram-bot[job-queue]
requests
beautifulsoup4
pytest
